Rewrite free server creation backend
