+++ app/Models/Billing/BillingException.php
@@ -15,6 +15,7 @@ use Illuminate\Database\Eloquent\Factories\HasFactory;
  * @property string $exception_type
  * @property \Carbon\Carbon $created_at
  * @property \Carbon\Carbon $updated_at
+ * @method static create(array $array)
  */
 class BillingException extends Model
 {
