+++ app/Services/Billing/CreateOrderService.php
@@ -11,7 +11,7 @@ class CreateOrderService
     /**
      * Process the creation of an order.
      */
-    public function create(?string $transaction_id, User $user, Product $product, ?string $status, ?string $type): Order
+    public function create(?string $transaction_id, User $user, Product $product, ?string $status, ?bool $is_new_order): Order
     {
         $order = new Order();
         $uuid = uuid_create();
@@ -23,7 +23,7 @@ class CreateOrderService
         $order->total = $product->price ?? 0;
         $order->status = $status ?? Order::STATUS_EXPIRED;
         $order->product_id = $product->id;
-        $order->type = $type;
+        $order->type = $is_new_order ? Order::TYPE_NEW : Order::TYPE_RENEWAL;
 
         $order->saveOrFail();
